+++ control-routing/src/main/java/fi/nls/oskari/control/routing/RoutingHandler.java
@@ -66,7 +66,7 @@ public class RoutingHandler extends ActionHandler {
 
         routeparams.setSrs(params.getHttpParam(PARAM_SRS));
         routeparams.setLang(params.getHttpParam(PARAM_LANGUAGE));
-        routeparams.setMaxWalkDistance(ConversionHelper.getLong(params.getHttpParam(PARAM_MAX_WALK_DISTANCE, PropertyUtil.get("routing.default.maxwalkdistance")), 1000));
+        routeparams.setMaxWalkDistance(ConversionHelper.getLong(params.getHttpParam(PARAM_MAX_WALK_DISTANCE, PropertyUtil.get("routing.default.maxwalkdistance")), 1000000));
         routeparams.setMode(params.getHttpParam(PARAM_MODE, PropertyUtil.get("routing.default.mode")));
 
         RouteResponse result = service.getRoute(routeparams);